Particle Effects
Added a explosion effect (obviously not final) that we can use. I think
the style suits the games aesthetics, will just need to modify
parameters if anything.
Also added a particle spawner and killer script. The killer script takes
a float and will destroy the game object after the float amount of
seconds. The spawner atm will only spawn in testing mode. But will
eventually have a public function to spawn the particle which will then
be called from other scripts when required.